+++ src/main/webapp/resources/js/components/icons/Icons.jsx
@@ -44,6 +44,8 @@ import {
   TeamOutlined,
   UnorderedListOutlined,
   UserOutlined,
+  UploadOutlined,
+  CalendarTwoTone,
 } from "@ant-design/icons";
 import { blue6 } from "../../styles/colors";
 
@@ -164,6 +166,10 @@ export const IconMembers = ({ ...props }) => <TeamOutlined {...props} />;
 
 export const IconUser = ({ ...props }) => <UserOutlined {...props} />;
 
+export const IconCalendarTwoTone = ({ ...props }) => (
+  <CalendarTwoTone {...props} />
+);
+
 /*
 File Type Icons
  */
@@ -193,3 +199,5 @@ export const IconSortDescending = ({ ...props }) => (
 export const IconCheckCircle = ({ ...props }) => (
   <CheckCircleOutlined {...props} />
 );
+
+export const IconFileUpload = ({ ...props }) => <UploadOutlined {...props} />;